Fix: Improve error handling in ImportAndReRunApply function by checking for errors before processing existing resources

+++ deploy/scripts/helpers/script_helpers.sh
@@ -571,16 +571,20 @@ function ImportAndReRunApply {
 					echo -e "${cyan}Terraform apply:                       succeeded$reset_formatting"
 					echo ""
 				fi
-				existing=$(jq 'select(."@level" == "error") | {address: .diagnostic.address, summary: .diagnostic.summary} | select(.summary | startswith("A resource with the ID"))' "$fileName")
-				if [[ -n ${existing} ]]; then
-					return_value=0
-				else
-					rm "$fileName"
+				errors_occurred=$(jq 'select(."@level" == "error") | length' "$fileName")
+				if [[ -n $errors_occurred ]]; then
+					existing=$(jq 'select(."@level" == "error") | {address: .diagnostic.address, summary: .diagnostic.summary} | select(.summary | startswith("A resource with the ID"))' "$fileName")
+					if [[ -n ${existing} ]]; then
+						return_value=0
+					else
+						rm "$fileName"
+					fi
 				fi
 			else
 				echo ""
 				echo -e "${cyan}No resources to import$reset_formatting"
 				echo ""
+				rm "$fileName"
 				return_value=1
 			fi
 		fi
